+++ WireNetwork/Sources/WireNetwork/APIs/Rest/ConversationsAPI/ConversationsAPIV8.swift
@@ -307,17 +307,19 @@ struct CreateGroupConversationParametersV8: Encodable {
 }
 
 extension ConversationGroupType {
+
     func toV8() -> ConversationGroupTypeV8? {
         switch self {
         case .group:
-            .group
+                .group
         case .channel:
-            .channel
+                .channel
         case .meeting:
             // Meeting was introduced in v15.
             nil
         }
     }
+
 }
 
 // MARK: - Decodables
